+++ packages/dso-toolkit/src/components/toggletip/toggletip.args.ts
@@ -1,22 +1,42 @@
 import { ArgTypes } from "@storybook/types";
 
-import { tooltipPositions } from "../tooltip/tooltip.models.js";
+import { BadgeStatus, badgeArgTypes } from "../badge";
+import { iconArgTypes } from "../icon";
 
-import { Toggletip } from "./toggletip.models.js";
+import { Toggletip, TooltipPosition } from "./toggletip.models.js";
 
 export interface ToggletipArgs {
-  position: "top" | "right" | "bottom" | "left";
+  mode?: "toggle" | "secondary" | "badge" | "icon";
+  position?: TooltipPosition;
+  strategy?: "absolute" | "fixed";
   small?: boolean;
   label?: string;
-  secondary?: boolean;
+  badgeStatus?: BadgeStatus;
+  icon?: string;
+  iconActive?: string;
 }
 
 export const toggletipArgTypes: ArgTypes<ToggletipArgs> = {
-  position: {
-    options: tooltipPositions,
+  mode: {
+    options: ["toggle", "secondary", "badge", "icon"],
     control: {
       type: "select",
     },
+    defaultValue: "toggle",
+  },
+  position: {
+    options: ["top", "right", "bottom", "left"],
+    control: {
+      type: "select",
+    },
+    defaultValue: "right",
+  },
+  strategy: {
+    options: ["absolute", "fixed"],
+    control: {
+      type: "select",
+    },
+    defaultValue: "absolute",
   },
   small: {
     control: {
@@ -28,9 +48,25 @@ export const toggletipArgTypes: ArgTypes<ToggletipArgs> = {
       type: "text",
     },
   },
-  secondary: {
-    control: {
-      type: "boolean",
+  badgeStatus: {
+    ...badgeArgTypes.status,
+    if: {
+      arg: "mode",
+      eq: "badge",
+    },
+  },
+  icon: {
+    ...iconArgTypes.icon,
+    if: {
+      arg: "mode",
+      eq: "icon",
+    },
+  },
+  iconActive: {
+    ...iconArgTypes.icon,
+    if: {
+      arg: "mode",
+      eq: "icon",
     },
   },
 };
@@ -41,9 +77,13 @@ export function toggletipArgsMapper<TemplateFnReturnType>(
 ): Toggletip<TemplateFnReturnType> {
   return {
     children,
+    mode: a.mode,
     position: a.position,
+    strategy: a.strategy,
     small: a.small,
     label: a.label,
-    secondary: a.secondary,
+    badgeStatus: a.badgeStatus,
+    icon: a.icon,
+    iconActive: a.iconActive,
   };
 }