+++ storybook/cypress/e2e/toggletip.cy.ts
@@ -8,25 +8,18 @@ describe("Toggletip", () => {
     cy.dsoCheckA11y("dso-toggletip.hydrated");
   });
 
-  function prepareComponent() {
+  it("should focus on button after escape", () => {
     cy.get("dso-toggletip.hydrated")
       .as("dsoToggletip")
       .shadow()
       .find("dso-info-button")
-      .as("dsoInfoButton")
       .shadow()
       .find("button")
       .as("dsoButton")
       .get("@dsoToggletip")
       .shadow()
-      .find("dso-tooltip")
+      .find(".tooltip")
       .as("dsoTooltip");
-  }
-
-  it("should focus on button after escape", () => {
-    prepareComponent();
-
-    cy.get("@dsoToggletip").matchImageSnapshot();
 
     cy.get("@dsoButton")
       .click()
@@ -39,4 +32,35 @@ describe("Toggletip", () => {
       .get("@dsoButton")
       .should("have.focus");
   });
+
+  const modes = ["toggle", "secondary", "badge", "icon"];
+
+  for (const mode of modes) {
+    it(`should show toggletip in ${mode} mode`, () => {
+      cy.visit(`http://localhost:45000/iframe.html?id=core-toggletip--toggletip&args=mode:${mode}`);
+
+      cy.get("dso-toggletip.hydrated").should("be.visible").matchImageSnapshot();
+    });
+  }
+
+  const positions = ["top", "right", "bottom", "left"];
+
+  for (const position of positions) {
+    it(`should show tooltip at ${position} position`, () => {
+      cy.viewport(750, 500).visit(
+        `http://localhost:45000/iframe.html?id=core-toggletip--toggletip&args=position:${position}`,
+      );
+
+      cy.get("dso-toggletip.hydrated")
+        .should("be.visible")
+        .shadow()
+        .find("dso-info-button")
+        .shadow()
+        .find("button")
+        .click();
+
+      // Even wachten tot de verschijntransistie van 150ms van het functional component tooltip gereed is
+      cy.wait(200).matchImageSnapshot();
+    });
+  }
 });